backend/pyproject.toml
[tool.ruff]
target-version = "py312" # Always generate Python 3.12-compatible code.
line-length = 88
indent-width = 4
lint.select = ["E", "F", "W", "B", "UP", "I", "C4", "SIM"]
# "ANN" checks for missing type annotations (currently removed)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.pyright]
typeCheckingMode = "strict"
reportOptionalSubscript = "error"
reportOptionalMemberAccess = "warning"
reportUnusedImport = "warning"
reportPrivateImportUsage = "none"

# Build backend
[build-system]
requires = ["setuptools>=80.9.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mbse-maturity"
version = "0.1.0"
description = "AI-assisted MBSE model maturity evaluator"
requires-python = ">=3.12,<3.13"


# Runtime dependencies
dependencies = [
  "fastapi==0.120.0",                     # Web framework
  "uvicorn[standard]==0.38.0",            # ASGI server
  "mcp>=1.21.2",                          # Model Conversion and Processing
  "python-multipart==0.0.20",             # Form data parsing for FastAPI
  "pydantic>=2",                          # Data validation and settings management
  "pydantic-settings>=2",                 # Settings management
  "defusedxml>=0.7.1",                    # Secure XML parsing
  "httpx>=0.28.1",                        # HTTP client
  "neo4j>=6.0.3",                         # Neo4j database driver
  "python-dotenv>=1.2.1",                # Environment variable management
  "openai>=2.8.1",                       # OpenAI API client
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.2",                      # Testing framework
  "pytest-cov>=7",                      # Test coverage
  "jsonschema>=4.25.1",                 # JSON Schema validation
]

# Package discovery: ship only the app code
[tool.setuptools.packages.find]
where = [""]                          # Search from the root  
include = ["app*"]                    # Include only packages starting with "app"   

# Uvicorn CLI defaults
[tool.uvicorn]
factory = false